MOSIP-21494 added join named native query

+++ mosip-compliance-toolkit/src/main/java/io/mosip/compliance/toolkit/dto/ProjectDto.java
@@ -17,7 +17,7 @@ public class ProjectDto {
 	
 	private String projectType;
 	
-	private int collectionsCount;
+	private long collectionsCount;
 	
 	private LocalDateTime crDate;
